requirements.txt
httpx>=0.27.0
PyYAML>=6.0.1
openai>=1.40.0
pytest>=8.0.0
